+++ packages/bfx-containers/src/containers/Grid/items/grid.balances.js
@@ -6,10 +6,10 @@ export default {
   id: 'balances',
   component: Balances,
   defaults: {
-    h: 12.8,
-    w: 4,
-    minW: 3,
     minH: 12,
+    h: 12.8,
+    minW: 3,
+    w: 4,
     static: false,
   },
   title: () => i18n.t('grid:balances'),