+++ packages/bfx-containers/src/containers/Trades/Trades.container.js
@@ -24,6 +24,7 @@ const TradesContainer = (props) => {
 
   const {
     isWSConnected,
+    isAuthenticated,
     symbol,
     dispatch,
   } = useCommonBfxData(baseCcy, quoteCcy)
@@ -37,8 +38,10 @@ const TradesContainer = (props) => {
   const minOrderSize = useSelector(state => getSymbolMinOrderSize(state, symbol))
 
   useEffect(() => {
-    dispatch(fetchAuthTradesHistory({ symbol }))
-  }, [dispatch, symbol])
+    if (isAuthenticated) {
+      dispatch(fetchAuthTradesHistory({ symbol }))
+    }
+  }, [dispatch, isAuthenticated, symbol])
 
   // resubscribe on symbol change
   useEffect(() => {