+++ packages/bfx-containers/src/redux/reducers/currencies.reducer.spec.js
@@ -1,0 +1,22 @@
+import types from '../constants/currencies.constants'
+import currenciesReducer, { INITIAL_STATE } from './currencies.reducer'
+import { newState, payload } from './currencies.reducer.data_for_test'
+
+describe('REDUCER: currencies', () => {
+  describe('action: REQUEST_CURRENCIES_INFO_SUCCESS', () => {
+    it('empty payload', () => {
+      const action = {
+        type: types.REQUEST_CURRENCIES_INFO_SUCCESS,
+        payload: [],
+      }
+      expect(currenciesReducer(INITIAL_STATE, action)).toEqual(INITIAL_STATE)
+    })
+    it('full payload (from real api response)', () => {
+      const action = {
+        type: types.REQUEST_CURRENCIES_INFO_SUCCESS,
+        payload,
+      }
+      expect(currenciesReducer(INITIAL_STATE, action)).toEqual(newState)
+    })
+  })
+})