+++ CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 
 
+## v0.12.0 (2021-12-20)
+
+#### :rocket: Feature
+* `bfx-containers`, `core`, `storybook`
+  * [#149](https://github.com/bitfinexcom/ufx-ui/pull/149) (feature) login with api creds ([@bhoomij](https://github.com/bhoomij))
+
+#### :dizzy: Improvement
+* `bfx-containers`
+  * [#147](https://github.com/bitfinexcom/ufx-ui/pull/147) (improvement) unsubscribe prev symbol related channels on ticker change ([@bhoomij](https://github.com/bhoomij))
+
+#### :bug: Bug Fix
+* `utils`
+  * [#152](https://github.com/bitfinexcom/ufx-ui/pull/152) (fix) decimals when price < 1 ([@bhoomij](https://github.com/bhoomij))
+* `bfx-containers`, `core`, `storybook`
+  * [#151](https://github.com/bitfinexcom/ufx-ui/pull/151) (fix) console error/warning ([@bhoomij](https://github.com/bhoomij))
+* Other
+  * [#148](https://github.com/bitfinexcom/ufx-ui/pull/148) (fix) line-break-style lint error on windows ([@bhoomij](https://github.com/bhoomij))
+
+#### :art: Refactor
+* `bfx-containers`, `core`, `storybook`
+  * [#150](https://github.com/bitfinexcom/ufx-ui/pull/150) (refactor) movements and paper trading improvements ([@bhoomij](https://github.com/bhoomij))
+
+#### Committers: 1
+- Bhoomi Joshi ([@bhoomij](https://github.com/bhoomij))
+
+
+
+## v0.11.5 (2021-11-23)
+
+#### :dizzy: Improvement
+* `core`
+  * [#143](https://github.com/bitfinexcom/ufx-ui/pull/143) (improvement) Modal: added Enter button support ([@dmytroshch](https://github.com/dmytroshch))
+
+#### :house: Internal
+* `bfx-containers`
+  * [#144](https://github.com/bitfinexcom/ufx-ui/pull/144) (improvement) Redux reducers tests ([@dmytroshch](https://github.com/dmytroshch))
+
+#### Committers: 2
+- Dmytro Shcherbonos ([@dmytroshch](https://github.com/dmytroshch))
+
+
+
+
+## v0.11.4 (2021-11-15)
+
+#### :dizzy: Improvement
+* `bfx-containers`
+  * [#142](https://github.com/bitfinexcom/ufx-ui/pull/142) (improvement) Using environment for Chart component ([@Blaumaus](https://github.com/Blaumaus))
+
+#### :bug: Bug Fix
+* `bfx-containers`
+  * [#140](https://github.com/bitfinexcom/ufx-ui/pull/140) (fix) Chart: add support for env REACT_APP_UFX_CORS_PROXY_URL ([@bhoomij](https://github.com/bhoomij))
+
+#### Committers: 2
+- Andrii R. ([@Blaumaus](https://github.com/Blaumaus))
+- Bhoomi Joshi ([@bhoomij](https://github.com/bhoomij))
+
+
 ## v0.11.3 (2021-10-27)
 
 #### :bug: Bug Fix